Switch versioning to semver (backward compatible)
From now on, the game version should be in semver notation.
This is mainly needed because the build scripts require it to be consistent, but it also makes it easier to do more specific version comparison.
A conversion function is used to handle old version numbers that are in float notation.